+++ DRN.Test.Integration/Tests/Sample/Utils/SourceKnownIdUtilsTests.cs
@@ -1,7 +1,18 @@
+using DRN.Framework.SharedKernel.Attributes;
+using DRN.Framework.SharedKernel.Domain;
 using DRN.Framework.Utils.Ids;
 using DRN.Framework.Utils.Time;
 
 namespace DRN.Test.Integration.Tests.Sample.Utils;
+
+public readonly struct CustomIntegrationTestApp : IAppId
+{
+    public const byte Value = 7;
+    public static byte AppId => Value;
+}
+
+[EntityType<CustomIntegrationTestApp>(1)]
+public class CustomIntegrationTestEntity : SourceKnownEntity;
 
 public class SourceKnownIdUtilsTests
 {
@@ -36,7 +47,7 @@ public class SourceKnownIdUtilsTests
             .WithExecutionMode(ParallelExecutionMode.ForceParallelism)
             .Select((_, index) =>
             {
-                ids[index] = generator.Next<ISourceKnownIdUtils>();
+                ids[index] = generator.Next<CustomIntegrationTestEntity>();
                 return index;
             }).ToArray();
         await Task.Delay(TimeStampManager.PrecisionUnitInMsSafeDelay); // wait for tick boundary